pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "django-auth-kit"
version = "0.1.0"
description = "A Django authentication kit with GraphQL (Strawberry), JWT, OTP, and social login support."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Django Auth Kit Contributors"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "django>=4.2",
    "strawberry-graphql>=0.220.0",
    "strawberry-graphql-django>=0.40.0",
    "PyJWT>=2.8.0",
]

[project.optional-dependencies]
social = [
    "django-allauth>=64.0.0",
]
dev = [
    "pytest",
    "pytest-django",
    "pytest-cov",
]

[tool.setuptools.packages.find]
include = ["django_auth_kit*"]

[tool.pytest.ini_options]
pythonpath = ["."]
